+++ buchfink/cli.py
@@ -13,6 +13,7 @@ from typing import Dict, List, Optional, Tuple, Union
 
 import click
 import coloredlogs
+import pyqrcode
 import yaml
 from rotkehlchen.chain.ethereum.trades import AMMTrade
 from rotkehlchen.constants import ZERO
@@ -123,6 +124,12 @@ def list_(buchfink_db: BuchfinkDB, keyword, account_type, output):
                     if account.tags \
                     else ''
             click.echo(type_and_name + address + tags)
+        elif output == 'qrcode':
+            if account.address:
+                qr = pyqrcode.create(account.address)
+                click.echo(qr.terminal(quiet_zone=1))
+            else:
+                click.echo('Can not create qrcode for {0}'.format(account.name))
         else:
             click.echo('{0}'.format(getattr(account, output)))